Remove "DBus" and "Service" from D-Bus service name and object path
To not give the impression that having "DBus" and/or "Service" in
service and interface names is something that should be done. Names become
less specific ("Template" instead of "TemplateDBusService") but this is
just a template and it is better to try to not be misleading.

A nice result of this change is that the name of the object on the bus
(/com/luxoft/Template) actually matches with DBusService::Template.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>