Remove sub-classing of stub in top-level DBusService class
A service can expose multiple objects easier in this way by just
subclassing another stub in a private class and adding an instance
variable. Also, multiple inheritance should be avoided. Object composition
is a better approach.

It also hides the stub more in DBusService and DBusService itself becomes
less cluttered with method overrides etc.

"Service" is a bad name for the internal class. Comes from the name of
the D-Bus interface. Will be fixed in a following commit.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>