Add a couple of type aliases in src/{cli,daemon}/main.cpp
Shorter to spell out type for return value of Arguments::parse(). More
clear what is going on than when using auto.

Avoid namespace aliases. Do not want to encourage use of them. The idea is
also that not much work should be done in 'main.cpp' files, so "need a ns
alias since doing so much work in main.cpp and much to type" does not fly. ;)

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>