Add short description and section for how to use cmcli to README.md
Also mention ConnMan in list of dependencies, but indicate that it is not
meant to be a "hard" dependency.

When more extensive documentation is added in e.g. a doc/ directory, some
information in README.md should probably be moved there (e.g. just mention
that cmcli exists in README.md and link to relevant section in doc/).

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>