Fix critical "cmcli monitor" GSource removal warning
The following error message is printed without this change when pressing
CTRL-C to stop monitoring:

(cmcli monitor:31379): GLib-CRITICAL **: 09:24:42.038: Source ID 1 was not found when attempting to remove it

CommandMonitor::monitor_until_ctrl_c() did originally not have a
g_source_remove() in it and relied on return value of callback for removal.
g_source_remove() was added during review of b03d7d88018b29a81c86 and was
not tested before merge.

Could remove g_source_remove() but it is probably better to not rely on
return value of callback if main loop exits for some other reason. E.g. we
decide to dynamically create proxies in a future version and decide to
quit if proxy creation fails. Then SIGINT callback will not be triggered.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>