Remove throttling of WiFiAccessPoints property in DBusService
Added to limit property changed signal emission for when many access points are
added/removed in less than 1s. Think it was added before the ADDED_ALL and
REMOVED_ALL events were added, in which case ADDED_ONE or REMOVED_ONE would
be emitted many times in a very short time when Wi-Fi was enabled/disabled.

Have not seen a problem with changing the property often but have on the
other hand seen problems in cmcli when this throttling is in place.
"cmcli wifi status" creates proxies for the paths in the WiFiAccessPoints
property and the throttling in DBusService made it more likely that a
proxy was used for an object that had been destroyed (not possible to
print access point information). It can still happen even without
throttling but is less likely. Could add code for handling this in cmcli
but keep it simple for now since it is just meant to be used for debugging.

Remove DBusService::wifi_access_points_changed() since setting the property
in the manager is now trivial. Have to update TODO comment about sorting
paths accordingly.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>