Add Security property to WiFiAccessPoint D-Bus objects
And implement it for ConnMan backend. The Security property of services in
ConnMan is an array of strings only present for Wi-Fi services. See
doc/service-api.txt in the ConnMan repository for details. We are currently
not interested in whether WPS is exposed or not so the "wps" and
"wps_advertising" strings are ignored for now. (Should probably expose this
as a separate boolean property in the WiFiAccessPoint interface if we want
to add it in the future and not complicate it with an array like in
ConnMan.)

Just print the property as is when listing access points with cmcli. Should
perhaps convert it to properly up-cased strings when presenting them to the
user, but cmcli is just a tool meant for debugging/testing so leave as is.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>