Remove TODO for detecting if a ConnMan service is a hidden AP
When setting description_type in Credentials::Requested that can be used
to indicate to user what request is for.

If a ConnMan service has "Type" set to "wifi" and is missing a "Name"
property, it is a hidden WiFi network. See doc/service-api.txt in the
ConnMan repo. ConnManService::name() defaults to the empty string when
no "Name" property exists.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>